requirements.txt
markovify
nltk
newsapi-python
praw
pytrends
datasets
TTS
gradio>=3.50.0
python-dotenv
jupyter
nbconvert
nbformat>=5.7.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.25.1
